+++ components/ftp/src/actions/directories/index.ts
@@ -1,0 +1,2 @@
+import createDirectory from "./createDirectory";
+export default { createDirectory };