+++ components/ftp/src/actions/files/deleteFile.ts
@@ -1,23 +1,14 @@
-import { action, input, util } from "@prismatic-io/spectral";
-import { connect } from "../../client";
+import { action } from "@prismatic-io/spectral";
+import { createClient } from "../../client";
 import { deleteFileExamplePayload } from "../../examplePayloads";
-import { connection, verbose } from "../../inputs";
-const path = input({
-  label: "Path",
-  placeholder: "Enter file path",
-  type: "string",
-  required: true,
-  comments: "The full path of the file on the FTP server to delete.",
-  example: "/path/to/file.txt",
-  clean: util.types.toString,
-});
-const deleteFile = action({
+import { deleteFileInputs } from "../../inputs";
+export const deleteFile = action({
   display: {
     label: "Delete File",
     description: "Deletes a file from an FTP server.",
   },
-  perform: async (_context, { connection, verbose, path }) => {
-    const client = await connect(connection, verbose);
+  perform: async (context, { connection, path }) => {
+    const client = await createClient(connection, context.debug.enabled);
     try {
       await client.remove(path);
     } finally {
@@ -25,7 +16,7 @@ const deleteFile = action({
     }
     return null;
   },
-  inputs: { connection, verbose, path },
+  inputs: deleteFileInputs,
   examplePayload: deleteFileExamplePayload,
 });
 export default deleteFile;