+++ components/ftp/src/actions/files/listDirectory.ts
@@ -1,23 +1,14 @@
-import { action, input, util } from "@prismatic-io/spectral";
-import { connect } from "../../client";
+import { action } from "@prismatic-io/spectral";
+import { createClient } from "../../client";
 import { listDirectoryExamplePayload } from "../../examplePayloads";
-import { connection, verbose } from "../../inputs";
-const path = input({
-  label: "Path",
-  placeholder: "Enter directory path",
-  type: "string",
-  required: true,
-  comments: "The full path of the directory on the FTP server to list.",
-  example: "/path/to/directory",
-  clean: util.types.toString,
-});
-const listDirectory = action({
+import { listDirectoryInputs } from "../../inputs";
+export const listDirectory = action({
   display: {
     label: "List Directory",
-    description: "List the contents of a directory",
+    description: "Lists the contents of a directory.",
   },
-  perform: async (_context, { connection, verbose, path }) => {
-    const client = await connect(connection, verbose);
+  perform: async (context, { connection, path }) => {
+    const client = await createClient(connection, context.debug.enabled);
     try {
       await client.cd(path);
       const contents = await client.list();
@@ -32,7 +23,7 @@ const listDirectory = action({
       client.close();
     }
   },
-  inputs: { connection, verbose, path },
+  inputs: listDirectoryInputs,
   examplePayload: listDirectoryExamplePayload,
 });
 export default listDirectory;