+++ components/ftp/src/actions/files/moveFile.ts
@@ -1,35 +1,14 @@
-import { action, input, util } from "@prismatic-io/spectral";
-import { connect } from "../../client";
+import { action } from "@prismatic-io/spectral";
+import { createClient } from "../../client";
 import { moveFileExamplePayload } from "../../examplePayloads";
-import { connection, verbose } from "../../inputs";
-const sourcePath = input({
-  label: "Source Path",
-  placeholder: "Enter source file path",
-  type: "string",
-  required: true,
-  comments: "The current path of the file on the FTP server to move.",
-  example: "/my/starting/path.txt",
-  clean: util.types.toString,
-});
-const destinationPath = input({
-  label: "Destination Path",
-  placeholder: "Enter destination file path",
-  type: "string",
-  required: true,
-  comments: "The new path where the file will be moved on the FTP server.",
-  example: "/my/destination/path.txt",
-  clean: util.types.toString,
-});
+import { moveFileInputs } from "../../inputs";
 export const moveFile = action({
   display: {
     label: "Move File",
     description: "Moves a file on an FTP server.",
   },
-  perform: async (
-    _context,
-    { connection, verbose, sourcePath, destinationPath },
-  ) => {
-    const client = await connect(connection, verbose);
+  perform: async (context, { connection, sourcePath, destinationPath }) => {
+    const client = await createClient(connection, context.debug.enabled);
     try {
       await client.rename(sourcePath, destinationPath);
     } finally {
@@ -37,12 +16,7 @@ export const moveFile = action({
     }
     return null;
   },
-  inputs: {
-    connection,
-    verbose,
-    sourcePath,
-    destinationPath,
-  },
+  inputs: moveFileInputs,
   examplePayload: moveFileExamplePayload,
 });
 export default moveFile;