+++ components/ftp/src/inputs/directories.ts
@@ -1,0 +1,6 @@
+import { connection, path, recursive } from "./common";
+export const createDirectoryInputs = {
+  connection,
+  path,
+  recursive,
+};