+++ components/ftp/src/inputs/triggers.ts
@@ -1,0 +1,7 @@
+import { connection, includeSubdirectories, path, pattern } from "./common";
+export const newOrModifiedFilesInputs = {
+  connection,
+  path,
+  pattern,
+  includeSubdirectories,
+};