+++ components/ftp/src/types/index.ts
@@ -1,13 +1,2 @@
-export interface FileMetadata {
-  size: number;
-  modifiedAt: number;
-}
-export type FileMap = Record<string, FileMetadata>;
-export interface PollingState {
-  fileMap: FileMap;
-}
-export interface FileEntry {
-  path: string;
-  size: number;
-  modifiedAt: number;
-}
+export * from "./files";
+export * from "./triggers";