+++ components/ftp/src/util/clean.ts
@@ -1,0 +1,5 @@
+import { util } from "@prismatic-io/spectral";
+export const cleanString = (value: unknown): string | undefined => {
+  const str = util.types.toString(value);
+  return str ? str : undefined;
+};